refactor: binding-friendly API signatures, type renames, and derives
- Rename types::OutputFormat → ExtractionMode (resolves name conflict
  with core::config::OutputFormat used for content format)
- Rename core config TokenReductionConfig → TokenReductionOptions
  (resolves conflict with text::TokenReductionConfig)
- Re-export text::TokenReductionConfig at crate root
- API signatures: &[&str] → &[String], &mut → return values,
  serde_json::Value → &str, char → u8, Path params → &str
- Add Default to: OcrElement, OcrConfidence, NodeId, ElementId,
  AnnotationKind, NodeContent, OcrBoundingGeometry, ErrorResponse,
  NativeTextStats, EmbeddingModelType, FormatMetadata
- Add Serialize/Deserialize to: BoundingBox (pdf), HierarchyLevel,
  RecognizedTable
- Exclude 30+ internal types and 20+ internal functions from
  alef binding generation
- Update WASM features to enable all non-ort features

+++ crates/kreuzberg/src/extractors/rtf/parser.rs
@@ -1192,7 +1192,7 @@ pub fn extract_text_from_rtf(
                             // Capture hex-encoded chars in footnote buffer even when skipping
                             if in_footnote
                                 && let (Some(h1), Some(h2)) = (hex1, hex2)
-                                && let Some(byte) = parse_hex_byte(h1, h2)
+                                && let Some(byte) = parse_hex_byte(h1 as u8, h2 as u8)
                             {
                                 footnote_buf.push(decode_windows_1252(byte));
                             }
@@ -1204,7 +1204,7 @@ pub fn extract_text_from_rtf(
                                 continue;
                             }
                             if let (Some(h1), Some(h2)) = (hex1, hex2)
-                                && let Some(byte) = parse_hex_byte(h1, h2)
+                                && let Some(byte) = parse_hex_byte(h1 as u8, h2 as u8)
                             {
                                 let decoded = decode_windows_1252(byte);
                                 if let Some(state) = table_state.as_mut()
